+++ app/src/ose/kotlin/com/davx5/ose/di/OseColorSchemesModule.kt
@@ -2,12 +2,12 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid.di
+package com.davx5.ose.di
 
 import androidx.compose.material3.ColorScheme
 import at.bitfire.davdroid.di.scope.DarkColorScheme
 import at.bitfire.davdroid.di.scope.LightColorScheme
-import at.bitfire.davdroid.ui.OseTheme
+import com.davx5.ose.ui.OseTheme
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
