+++ app/src/ose/kotlin/com/davx5/ose/ui/intro/OseIntroPageFactory.kt
@@ -2,8 +2,15 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid.ui.intro
+package com.davx5.ose.ui.intro
 
+import at.bitfire.davdroid.ui.intro.BackupsPage
+import at.bitfire.davdroid.ui.intro.BatteryOptimizationsPage
+import at.bitfire.davdroid.ui.intro.IntroPageFactory
+import at.bitfire.davdroid.ui.intro.OpenSourcePage
+import at.bitfire.davdroid.ui.intro.PermissionsIntroPage
+import at.bitfire.davdroid.ui.intro.TasksIntroPage
+import at.bitfire.davdroid.ui.intro.WelcomePage
 import javax.inject.Inject
 
 class OseIntroPageFactory @Inject constructor(
